(#1406) Move procedures to a designated package
