+++ src/main/java/org/cactoos/BiProc.java
@@ -23,13 +23,16 @@
  */
 package org.cactoos;
 
+import org.cactoos.proc.IoCheckedBiProc;
+import org.cactoos.proc.UncheckedBiProc;
+
 /**
  * Proc that accepts two arguments.
  *
  * <p>If you don't want to have any checked exceptions being thrown
  * out of your {@link BiProc}, you can use
- * {@link org.cactoos.func.UncheckedBiProc} decorator. Also
- * you may try {@link org.cactoos.func.IoCheckedBiProc}.</p>
+ * {@link UncheckedBiProc} decorator. Also
+ * you may try {@link IoCheckedBiProc}.</p>
  *
  * <p>There is no thread-safety guarantee.
  *
