+++ src/main/java/org/cactoos/io/TempFolder.java
@@ -32,10 +32,10 @@ import java.nio.file.Paths;
 import java.util.Comparator;
 import org.cactoos.Scalar;
 import org.cactoos.Text;
-import org.cactoos.func.ForEach;
-import org.cactoos.func.IoCheckedProc;
-import org.cactoos.func.ProcOf;
 import org.cactoos.iterable.Sorted;
+import org.cactoos.proc.ForEach;
+import org.cactoos.proc.IoCheckedProc;
+import org.cactoos.proc.ProcOf;
 import org.cactoos.scalar.IoChecked;
 import org.cactoos.scalar.Sticky;
 import org.cactoos.text.Joined;
