+++ src/main/java/org/cactoos/proc/package-info.java
@@ -21,45 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.func;
-
-import org.cactoos.Func;
-import org.cactoos.Proc;
 
 /**
- * Proc that runs repeatedly for a number of times.
+ * Procedures.
  *
- * @param <X> Type of input
- * @since 0.49.2
- * @todo #1277:30m Split the entire `org.cactoos.func` package
- *  into `org.cactoos.func`, `org.cactoos.proc`
- *  and `org.cactoos.callable`,
- *  with each one holding their related decorators, classes and interfaces.
+ * @since 0.47
  */
-public final class RepeatedProc<X> implements Proc<X> {
-    /**
-     * The repeated func.
-     */
-    private final Func<X, Boolean> func;
-
-    /**
-     * Ctor.
-     *
-     * <p>If {@code count} is equal or less than zero {@link #exec(Object)}
-     * will return an exception.</p>
-     *
-     * @param prc Proc to repeat.
-     * @param count How many times.
-     */
-    public RepeatedProc(final Proc<X> prc, final int count) {
-        this.func = new Repeated<>(
-            new FuncOf<>(prc, true),
-            count
-        );
-    }
-
-    @Override
-    public void exec(final X input) throws Exception {
-        this.func.apply(input);
-    }
-}
+package org.cactoos.proc;
