+++ src/test/java/org/cactoos/func/package-info.java
@@ -23,7 +23,7 @@
  */
 
 /**
- * Functions and procedures, tests.
+ * Functions, tests.
  *
  * @since 0.2
  */
