+++ src/test/java/org/cactoos/io/TempFolderTest.java
@@ -26,9 +26,9 @@ package org.cactoos.io;
 import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import org.cactoos.func.ForEach;
-import org.cactoos.func.ProcOf;
 import org.cactoos.iterable.IterableOf;
+import org.cactoos.proc.ForEach;
+import org.cactoos.proc.ProcOf;
 import org.cactoos.text.Randomized;
 import org.hamcrest.core.IsNot;
 import org.junit.Test;
