+++ src/test/java/org/cactoos/proc/BiProcNoNullsTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.func;
+package org.cactoos.proc;
 
 import java.util.concurrent.atomic.AtomicInteger;
 import org.hamcrest.core.IsEqual;
