+++ src/test/java/org/cactoos/proc/IoCheckedBiProcTest.java
@@ -21,10 +21,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.func;
+package org.cactoos.proc;
 
 import java.io.IOException;
 import java.util.concurrent.atomic.AtomicInteger;
+import org.cactoos.func.IoCheckedBiFunc;
 import org.hamcrest.core.IsEqual;
 import org.junit.Test;
 import org.llorllale.cactoos.matchers.Assertion;
